# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movies API",
        "description": "Movies API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Movies API"
        }
    ],
    "paths": {
...
        "/api/movies": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get all movies",
                "description": "Returns a list of movies. This endpoint requires a valid Bearer token in the Authorization header.",
                "operationId": "getMovies",
                "responses": {
                    "200": {
                        "description": "A list with movies"
                    },
                    "401": {
                        "description": "Unauthorized. The request must include a valid Bearer token in the Authorization header."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/movies/{movie_id}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get a specific movie",
                "description": "Returns a specific movie by ID.",
                "operationId": "getMovieById",
                "parameters": [
                    {
                        "name": "movie_id",
                        "in": "path",
                        "description": "ID of the movie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The movie details"
                    },
                    "401": {
                        "description": "Unauthorized. The request must include a valid Bearer token in the Authorization header."
                    },
                    "404": {
                        "description": "Movie not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/movies/{movie_id}/genres": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get genres associated with a specific movie",
                "description": "Returns a list of genres associated with a specific movie.",
                "operationId": "getMovieGenres",
                "parameters": [
                    {
                        "name": "movie_id",
                        "in": "path",
                        "description": "ID of the movie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list with genres"
                    },
                    "401": {
                        "description": "Unauthorized. The request must include a valid Bearer token in the Authorization header."
                    },
                    "404": {
                        "description": "Movie not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/movies/{movie_id}/rating": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get ratings associated with a specific movie",
                "description": "Returns a list of ratings associated with a specific movie.",
                "operationId": "getMovieRating",
                "parameters": [
                    {
                        "name": "movie_id",
                        "in": "path",
                        "description": "ID of the movie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list with ratings"
                    },
                    "401": {
                        "description": "Unauthorized. The request must include a valid Bearer token in the Authorization header."
                    },
                    "404": {
                        "description": "Movie not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/movies/{movie_id}/directors": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get directors associated with a specific movie",
                "description": "Returns a list of directors associated with a specific movie.",
                "operationId": "getMovieDirectors",
                "parameters": [
                    {
                        "name": "movie_id",
                        "in": "path",
                        "description": "ID of the movie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list with directors"
                    },
                    "401": {
                        "description": "Unauthorized. The request must include a valid Bearer token in the Authorization header."
                    },
                    "404": {
                        "description": "Movie not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/movies/{movie_id}/actors": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get actors associated with a specific movie",
                "description": "Returns a list of actors associated with a specific movie.",
                "operationId": "getMovieActors",
                "parameters": [
                    {
                        "name": "movie_id",
                        "in": "path",
                        "description": "ID of the movie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list with actors"
                    },
                    "401": {
                        "description": "Unauthorized. The request must include a valid Bearer token in the Authorization header."
                    },
                    "404": {
                        "description": "Movie not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the token without the 'Bearer' prefix, it is added automatically. We recommend you first log in and get a token.",
                "name": "oauth2",
                "in": "header",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "oauth/token",
                        "scopes": {}
                    }
                },
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}